fix metamask page
