+++ src/stylesheets/components/slices/_blog.scss
