+++ src/stylesheets/components/slices/_meta-mask.scss
@@ -51,12 +51,13 @@
   }
   &__login:disabled {
     padding: 5px;
-    margin: 12px;
     border-radius: 10px;
     color: white;
     border: 1px solid white;
     background-color: #cecece;
     font-size: 20px;
+        margin:15px;
+
   }
   &__logout{
     padding: 15px;
@@ -64,6 +65,18 @@
     color: white;
     border: 1px solid white;
     background-color: #ff5e5e;
+    font-weight: bold;
+    font-size: 22px;
+        margin:15px;
+
+  }
+  &__btn{
+    padding: 15px;
+    margin:15px;
+    border-radius: 10px;
+    color: black;
+    border: 1px solid black;
+    background-color: white;
     font-weight: bold;
     font-size: 22px;
   }
