+++ src/stylesheets/components/slices/_my-projects.scss
